Flyttet search klasser til libs-dto
